fixes the issue #401: k8s secret not working
