+++ components/micro-gateway-cli/src/main/resources/templates/kubernetesSecret.mustache
@@ -1,9 +1,14 @@
 {{#if containerConfig.kubernetes.kubernetesSecret.enable}}
+{{#containerConfig.kubernetes.kubernetesSecret}}
 @kubernetes:Secret {
-    {{#if containerConfig.kubernetes.kubernetesSecret.name}}
-    name:"{{containerConfig.kubernetes.kubernetesSecret.name}}"{{/if}}{{#if containerConfig.kubernetes.kubernetesSecret.mountPath}},
-    mountPath:"{{containerConfig.kubernetes.kubernetesSecret.mountPath}}"{{/if}}{{#if containerConfig.kubernetes.kubernetesSecret.readOnly}},
-    readOnly:"{{containerConfig.kubernetes.kubernetesSecret.readOnly}}"{{/if}}{{#if containerConfig.kubernetes.kubernetesSecret.data}},
-    data:"{{containerConfig.kubernetes.kubernetesSecret.accessMode}}"{{/if}}
-}
+    {{#if secrets}}
+        secrets: [ {{#secrets}}
+            { {{#name}}
+                name:"{{name}}",{{/name}}{{#readOnly}}
+                readOnly:{{readOnly}},{{/readOnly}}{{#if data}}
+                data:[{{#data}}"{{.}}"{{#unless @last}},{{/unless}}{{/data}}],{{/if}}{{#mountPath}}
+                mountPath:"{{mountPath}}"{{/mountPath}}
+            }{{#unless @last}},{{/unless}}{{/secrets}}
+        ]{{/if}}
+}{{/containerConfig.kubernetes.kubernetesSecret}}
 {{/if}}
